--- INFO.txt
@@ -1,92 +1,0 @@
- INFO
-
- Versione Corrente = 1.6.2
- 
- Versione 1.6.2:
-   - Aggiunti labels per visualizzare il contenuto di AC in intero ed esadecimale
-
- Versione 1.6.1:
-   - (Core) Corretta addizione dopo aggiunta delle micro-istruzioni                                             
-
- Versione 1.6 :                                                         
-   - Creato il presente file (INFO.txt) per le informazioni degli aggiornamenti                    
-   - Corretto salvataggio file (non metteva l'estensione)               
-   - Se si apre un nuovo file, viene cancellato il contenuto attuale    
-     dell'editor e si carica il file                                    
-   - Non è più necessario premere reset per caricare un nuovo file.     
-     Load provvederà a cancellare il contenuto attuale della memoria    
-   - Correzioni varie alla chiusura della finestra dell'emulatore       
-   - (Core) Eliminato metodo statico elws (basta rstrip)                
-   - (Core) sostituito in binario 2**16 con 65536, visto che è una      
-     costante                                                           
-   - (Core) Eliminati extendpositive ed extendlabel che sono stati      
-     sostituiti da zfill                                                
-   - (Core) Aggiunto controllo se viene inserita o meno l'isruzione     
-     HLT                                                                
-   - (Core) Eliminati metodi statici hex2dec, bin2dec, dec2char e       
-     char2dec per utilizzare le funzioni standard del python
-   - (Core) Corretta ricerca dell'END a fine file.
-
- Versione 1.5 :                                                         
-   - (Core) Riscritto il core per adattarlo all'esecuzione delle        
-     singole microistruzioni                                            
-   - Riadattata interfaccia per l'esecuzione delle singole              
-     microistruioni                                                     
-
- Versione 1.4 :                                                         
-   - Aggiunto colore viola del label start quando incontra un break     
-   - Aggiunto messaggio di warning quando non si carica in modo         
-     corretto il codice. Si riporta la parola errata e si riconoscono   
-     gli errori di ORG ed END. Si riconoscono solo errori sintattici    
-   - Tolto next istruction di colore verde e messo ultima istruzione    
-     eseguita in verde                                                  
-   - Eliminata colonna indirizzi in decimale                            
-   - Corretto tasto start quando si ferma ai break                      
-   - Corretta ricerca label (eliminate spaziature superflue nei nomi)   
-
- Versione 1.34 :                                                        
-   - Corretto miniStep, non aggiornava i registri                       
-   - Aggiunto esempio MOLTIPLICAZIONE ed aggiornato mulIO               
-
- Versione 1.33 :                                                        
-   - Corretto setdelay per gestire il ritorno None, cioè quando         
-     l'utente preme cancel                                              
- Versione 1.32 :                                                        
-   - Corretto ricerca break nello status ram                            
-   - Corretto setnstep per gestire il ritorno None, cioè quando         
-     l'utente preme cancel                                              
-
- Versione 1.31 :                                                        
-   - Aggiornato chiusura delle finestre, ora sono entrambe collegate    
-     alla funzione exit della classe editor                             
-
- Versione 1.3 :                                                         
-   - Aggiunto pulsante break, per inserire o togliere un breakpoint     
-     nel codice da eseguire                                             
-   - Aggiunto pulsante continua, per riprendere l'esecuzione dopo       
-     un breakpoint                                                      
-   - (Core) Aggiunto metodo per inserire i breakpoints                  
-   - (Core) Modificato step per bloccarsi dopo l'esecuzione se          
-     incontra un breakpoint                                             
-
- Versione 1.2 :                                                         
-   - Suddivisione delle classi in più files (Walter Valentini)          
-   - Riorganizzazione degli import (Walter Valentini)                   
-   - Trasformazione in package python (Walter Valentini)                
-   - Rimozione degli spazi non necessari (Walter Valentini)             
-
- Versione 1.1 :                                                         
-   - Divisione finestra editor per il codice assembly e finestra per    
-     l'esecuzione del codice.                                           
-   - Ridimensionamento dei vari frame e riposizionamento degli oggetti  
-   - Corretti cerca MAR e cerca PC se ci sono più ORG nel file          
-     assembly                                                           
-   - Aggiunta funzione mini_step che esegue un singolo ciclo in base    
-     alle variabili F ed R dell'UC                                      
-   - (Core) Aggiunto ritorno carrello quando si pigia return            
-     senza immettere caratteri. Tolto output di INP (necessario         
-     solo per debugging)                                                
-   - (Core) Aggiornato OUT per stampare correttamente il ritorno        
-     carrello, ovvero il fine riga.                                     
-
- Versione 1.0 : Release Candidate
