+++ Version.txt
@@ -1,0 +1,15 @@
+Pypdp8wx Emulator
+
+Versione corrente : 2.0
+
+Cambiamente dalle precedenti versioni:
+
+V 2.0 : 
+ - Riscritto l'intero programma con librerie grafiche wxWidgets in python
+ - Adattato il core per il nuovo programma
+ - (Editor) Highlights delle keyword, dei numeri e degli operatori per il codice
+   assembly del pdp8 e conteggio delle linee
+ - (Emulatore RAM) Migliorata la visuale della memoria e aggiunta la possibilità
+   di selezionare una cella di memoria per visualizzare il contenuto in HEX
+   e la parte indirizzo in HEX
+ - Migliorie varie al codice
