--- pypdp8 - wxPython/__main__.py
@@ -1,20 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-'''
-Created on Jun 23, 2011
-
-@author: Mirco Tracolli
-'''
-
-from wxGUI import APP
-from sys import exit
-
-
-def run():
-    App = APP(False)
-
-    App.MainLoop()
-
-if __name__ == '__main__':
-    exit(run())
